+++ app/src/types.ts
@@ -146,7 +146,6 @@ export type PlaylistType = {
     type: string;
     picture: string;
   }[];
-
   lastItemAddedAt: string;
 };
 
@@ -237,6 +236,7 @@ export type ConfigTiddleType = {
     embed_lyrics: boolean;
     download_video: boolean;
     scan_path: string;
+    save_playlist_m3u: boolean;
   };
   cover: {
     save: boolean;
@@ -286,13 +286,20 @@ export type LogType = {
 // PROCESSING LIST
 
 export type QualityType = "low" | "normal" | "high" | "master";
+export type ContentType =
+  | "artist"
+  | "album"
+  | "track"
+  | "playlist"
+  | "video"
+  | "mix";
 
 export type ProcessingItemType = {
   id: string;
   artist: string;
   title: string;
   quality: QualityType;
-  type: "artist" | "album" | "track" | "playlist" | "video" | "mix";
+  type: ContentType;
   status: "queue" | "finished" | "beet" | "processing" | "error";
   url: string;
   loading: boolean;
@@ -304,4 +311,15 @@ export type ApiReturnType = {
   error: boolean;
   message: string;
   status?: number;
+};
+
+// SYNC LIST
+
+export type SyncItemType = {
+  id: string;
+  title: string;
+  url: string;
+  quality: QualityType;
+  lastUpdate?: string;
+  type: ContentType;
 };